Add files via upload
Shortened the delay between drawing cards in automatic 'play-bot' mode for the Dealer.  Change the card height to be 40% of the play window's size to prevent the cards from getting very large as the screen's view port size is reduced.  Added automatic scroll bars to the 'play window' if the played cards overflow the 'play window' box due smaller screen sizes.